backend/requirements.txt
Django==4.2.16
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.4.0
python-decouple==3.8
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
pytest==7.4.3
pytest-django==4.7.0
coverage==7.3.2
factory-boy==3.3.1